+++ test/data/match-path-data.ts
@@ -95,9 +95,7 @@ export const tests: ReadonlyArray<OneTest> = [
     existingFiles: [join("/root", "location", "mylib", "kalle.ts")],
     packageJson: { main: "./kalle.ts" },
     requestedModule: "lib/mylib",
-    expectedPath: removeExtension(
-      join("/root", "location", "mylib", "kalle.ts")
-    ),
+    expectedPath: join("/root", "location", "mylib", "kalle.ts"),
   },
   {
     name: "should resolve from main field in package.json (js)",
@@ -107,22 +105,7 @@ export const tests: ReadonlyArray<OneTest> = [
     packageJson: { main: "./kalle.js" },
     requestedModule: "lib/mylib.js",
     extensions: [".ts", ".js"],
-    expectedPath: removeExtension(
-      join("/root", "location", "mylib.js", "kalle.js")
-    ),
-  },
-  {
-    name:
-      "should resolve from main field in package.json and correctly remove file extension",
-    absoluteBaseUrl: "/root/",
-    paths: { "lib/*": ["location/*"] },
-    existingFiles: [join("/root", "location", "mylibjs", "kalle.js")],
-    packageJson: { main: "./kalle.js" },
-    extensions: [".ts", ".js"],
-    requestedModule: "lib/mylibjs",
-    expectedPath: removeExtension(
-      join("/root", "location", "mylibjs", "kalle.js")
-    ),
+    expectedPath: join("/root", "location", "mylib.js", "kalle.js"),
   },
   {
     name: "should resolve from list of fields by priority in package.json",
@@ -136,9 +119,7 @@ export const tests: ReadonlyArray<OneTest> = [
     ],
     extensions: [".ts", ".js"],
     requestedModule: "lib/mylibjs",
-    expectedPath: removeExtension(
-      join("/root", "location", "mylibjs", "browser.js")
-    ),
+    expectedPath: join("/root", "location", "mylibjs", "browser.js"),
   },
   {
     name: "should ignore field mappings to missing files in package.json",
@@ -152,9 +133,7 @@ export const tests: ReadonlyArray<OneTest> = [
       browser: "./nope.js",
     },
     extensions: [".ts", ".js"],
-    expectedPath: removeExtension(
-      join("/root", "location", "mylibjs", "kalle.js")
-    ),
+    expectedPath: join("/root", "location", "mylibjs", "kalle.js"),
   },
   {
     name: "should ignore advanced field mappings in package.json",
@@ -170,9 +149,7 @@ export const tests: ReadonlyArray<OneTest> = [
       browser: { mylibjs: "./browser.js", "./kalle.js": "./browser.js" },
     },
     extensions: [".ts", ".js"],
-    expectedPath: removeExtension(
-      join("/root", "location", "mylibjs", "kalle.js")
-    ),
+    expectedPath: join("/root", "location", "mylibjs", "kalle.js"),
   },
   {
     name: "should resolve to with the help of baseUrl when not explicitly set",
@@ -208,5 +185,16 @@ export const tests: ReadonlyArray<OneTest> = [
     existingFiles: [join("/root", "location", "mylib", "index.d.ts")],
     requestedModule: "lib/mylib",
     expectedPath: undefined,
+  },
+  {
+    name: "should resolve main file with cjs file extension",
+    absoluteBaseUrl: "/root/",
+    paths: {},
+    existingFiles: [join("/root", "mylib", "index.cjs")],
+    packageJson: {
+      main: "./index.cjs",
+    },
+    requestedModule: "mylib",
+    expectedPath: join("/root", "mylib", "index.cjs"),
   },
 ];